+++ motionit/.gitattributes
@@ -1,3 +1,3 @@
 /gradlew text eol=lf
-*.bat text eol=crlf
+*.bat text eol=lf
 *.jar binary
