+++ motionit/config/checkstyle/naver-checkstyle-suppressions.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<!--suppress XmlUnboundNsPrefix, XmlDefaultAttributeValue -->
+<!DOCTYPE suppressions PUBLIC
+        "-//Puppy Crawl//DTD Suppressions 1.1//EN"
+        "http://www.puppycrawl.com/dtds/suppressions_1_1.dtd">
+
+<suppressions>
+</suppressions>
