+++ motionit/src/main/java/com/back/motionit/MotionitApplication.java
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @EnableJpaAuditing
 public class MotionitApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(MotionitApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(MotionitApplication.class, args);
+	}
 
 }
