+++ motionit/src/main/java/com/back/motionit/global/jpa/entity/BaseEntity.java
@@ -1,5 +1,11 @@
 package com.back.motionit.global.jpa.entity;
 
+import java.time.LocalDateTime;
+
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
@@ -7,25 +13,20 @@ import jakarta.persistence.MappedSuperclass;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
 
 @MappedSuperclass
 @Getter
 @EntityListeners(AuditingEntityListener.class)
 public abstract class BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = jakarta.persistence.GenerationType.IDENTITY)
-    @Setter(AccessLevel.PROTECTED)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = jakarta.persistence.GenerationType.IDENTITY)
+	@Setter(AccessLevel.PROTECTED)
+	private Long id;
 
-    @CreatedDate
-    private LocalDateTime createDate;
+	@CreatedDate
+	private LocalDateTime createDate;
 
-    @LastModifiedDate
-    private LocalDateTime modifyDate;
+	@LastModifiedDate
+	private LocalDateTime modifyDate;
 }
