--- motionit/src/main/java/com/back/motionit/global/springDoc/SpringDoc.java
@@ -1,37 +1,0 @@
-package com.back.motionit.global.springDoc;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
-import io.swagger.v3.oas.annotations.info.Info;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import org.springdoc.core.models.GroupedOpenApi;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@OpenAPIDefinition(info = @Info(title = "API 서버", version = "beta", description = "API 서버 문서입니다."))
-@SecurityScheme(
-        name = "bearerAuth",
-        type = SecuritySchemeType.HTTP,
-        bearerFormat = "JWT",
-        scheme = "bearer"
-)
-public class SpringDoc {
-
-    @Bean
-    public GroupedOpenApi groupApiV1() {
-        return GroupedOpenApi.builder()
-                .group("apiV1")
-                .pathsToMatch("/api/v1/**")
-                .build();
-    }
-
-    @Bean
-    public GroupedOpenApi groupController() {
-        return GroupedOpenApi.builder()
-                .group("home")
-                .pathsToExclude("/api/**")
-                .build();
-    }
-
-}
