+++ motionit/src/main/java/com/back/motionit/global/webmvcconfig/WebMvcConfig.java
@@ -1,0 +1,18 @@
+package com.back.motionit.global.webmvcconfig;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+	@Override
+	public void addCorsMappings(CorsRegistry registry) {
+		registry.addMapping("/api/**")
+			.allowedOrigins("http://localhost:3000")
+			.allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS")
+			.allowedHeaders("*")
+			.allowCredentials(true);
+	}
+}
